Revert Guards.java to generated version, remove dsl-compiler.exe, updates to the Makefile, refactoring of the pregen dir structure and the newest dsl-clc.jar.
