+++ tpc/Makefile
@@ -32,7 +32,8 @@ cks_generator ?= java -jar lib/cks-tool.jar
 avro_cp ?= $(call mkcp,$(wildcard lib/avro-*.jar lib/jackson-*.jar lib/jopt-simple-*.jar lib/velocity-*.jar lib/slf4j-*.jar))
 avro_generator ?= java -cp "$(avro_cp)" org.apache.avro.tool.Main
 protostuff_compiler ?= $(wildcard lib/protostuff-compiler-*.jar)
-dslplatform_generator = java -jar lib/dsl-clc.jar
+dslplatform_compiler = java -jar lib/dsl-clc.jar
+dslplatform_dir = build/dslplatform
 
 # Default to 'no' because we might not have these tools available.
 Generate_protobuf ?= n
@@ -127,19 +128,19 @@ else ifneq ($(findstring .tcks,$(2)),)
 else ifneq ($(findstring .avro,$(2)),)
 	$(p1)$(avro_generator) compile schema $$< build/gen/$(2)
 else ifneq ($(findstring .dsl,$(2)),)
-	rm -rf dslplatform_client_tmp java_client
-	mkdir -p dslplatform_client_tmp
-	$(p1)$(dslplatform_generator) \
-		-temp=dslplatform_client_tmp \
+	$(p1)mkdir -p $(dslplatform_dir)/tmp
+	$(p1)$(dslplatform_compiler) \
+		-temp=$(dslplatform_dir)/tmp \
 		-dsl=schema \
 		-settings=manual-json,no-jackson \
 		-namespace=serializers.dslplatform \
 		-target=java_client \
 		-no-prompt \
-		-compiler=lib/dsl-compiler.exe \
+		-compiler \
+		-dependencies:java_client=$(dslplatform_dir)/dependencies \
+		-java_client=$(dslplatform_dir)/model.jar \
 		-download
-	cp -r dslplatform_client_tmp/java_client/serializers build/gen/$(2)
-	rm -rf dslplatform_client_tmp java_client generated-model-java.jar
+	$(p1)cp -r $(dslplatform_dir)/tmp/java_client/serializers build/gen/$(2)
 else
 endif
 
@@ -213,3 +214,4 @@ ifneq ($(wildcard build),)
 	# Delete: build/**
 	$(p1)-rm -r build
 endif
+
