--- tpc/pregen/media.dsl/dslplatform/media/Player.java
@@ -1,6 +1,0 @@
-package serializers.dslplatform.media;
-
-public enum Player {
-    JAVA,
-    FLASH
-}
