--- tpc/pregen/media.dsl/dslplatform/media/Size.java
@@ -1,6 +1,0 @@
-package serializers.dslplatform.media;
-
-public enum Size {
-    SMALL,
-    LARGE
-}
