+++ tpc/pregen/media.dsl/serializers/dslplatform/media/Player.java
@@ -1,0 +1,8 @@
+package serializers.dslplatform.media;
+
+
+
+public enum Player { 
+	JAVA,
+	FLASH
+}
