+++ tpc/pregen/media.dsl/serializers/dslplatform/media/Size.java
@@ -1,0 +1,8 @@
+package serializers.dslplatform.media;
+
+
+
+public enum Size { 
+	SMALL,
+	LARGE
+}
